:recycle: Refactor txt file detection (again)
This approach should hopefully cover most cases.
First we try to find a `_chat.txt` file which is a very common name.
Otherwise we search for `.txt` files with `chat` or `whatsapp` in the name, if we find any matches we sort them by name length and select the shortest one.